Aggiunti i service e cambiato tutti i long in int

+++ src/main/java/it/contrader/controller/AbstractController.java
@@ -34,7 +34,7 @@ public abstract class AbstractController <DTO>{
 	}
 	
 	@DeleteMapping("/delete")
-	public void delete(@RequestParam("id") long id) {
+	public void delete(@RequestParam("id") int id) {
 		service.delete(id);
 	}
 	
@@ -51,7 +51,7 @@ public abstract class AbstractController <DTO>{
 	}
 	
 	@GetMapping("/read")
-	public DTO read(long id) {
+	public DTO read(int id) {
 		return service.read(id);
 	}
 }
